+++ client-app/pages/sign-up.vue
@@ -43,7 +43,7 @@
           :error-message="errors.email"
         ></VcInput>
         <VcInput
-          v-if="registrationKind === RegistrationKind.organization"
+          v-if="registrationKind == RegistrationKind.organization"
           v-model="organizationName"
           class="mb-4"
           :label="$t('pages.sign_up.organization_name_label')"
@@ -110,7 +110,7 @@ import { useForm, useField } from "vee-validate";
 import * as yup from "yup";
 import { ref } from "vue";
 import { usePopup } from "@/shared/popup";
-import { IdentityResultType } from "@/xapi/graphql/types";
+import { AccountCreationResultType, IdentityErrorType } from "@/xapi/graphql/types";
 import { computed } from "@vue/reactivity";
 import { isObjectEmpty, trimString } from "@/core/utilities";
 import { useI18n } from "vue-i18n";
@@ -191,8 +191,8 @@ const commonErrors = ref<string[]>([]);
 const onSubmit = handleSubmit(async (data) => {
   commonErrors.value = [];
 
-  let result: IdentityResultType;
-  if (registrationKind.value === RegistrationKind.personal) {
+  let result: AccountCreationResultType;
+  if (registrationKind.value == RegistrationKind.personal) {
     result = await registerUser({
       email: trimString(data.email),
       firstName: trimString(data.firstName),
@@ -217,11 +217,11 @@ const onSubmit = handleSubmit(async (data) => {
     });
   } else {
     if (result.errors?.length) {
-      for (let i = 0; i < result.errors?.length; i++) {
-        const error = result.errors[i];
+      const identityErrors = convertToIdentityError(result.errors);
+      for (let error of identityErrors) {
         // TODO: Localize all messages on a front side
         // improve XAPI: pass error's parameters with error
-        switch (error?.code) {
+        switch (error.code) {
           case "password-too-weak":
           case "PasswordTooShort":
           case "PasswordRequiresLower":
@@ -233,34 +233,55 @@ const onSubmit = handleSubmit(async (data) => {
           case "InvalidPasswordHasherCompatibilityMode":
           case "InvalidPasswordHasherIterationCount":
             // t() is the workaround for the empty description.
-            setFieldError("password", error?.description || t("pages.sign_up.errors." + error.code));
+            setFieldError("password", error.description || t("pages.sign_up.errors." + error.code));
             break;
 
           case "repeat-password":
           case "PasswordMismatch":
-            setFieldError("confirmPassword", error?.description);
+            setFieldError("confirmPassword", error.description);
             break;
 
           case "DuplicateUserName":
           case "InvalidUserName":
           case "LoginAlreadyAssociated":
-            setFieldError("userName", error?.description);
+            setFieldError("userName", error.description);
             break;
 
           case "DuplicateEmail":
           case "InvalidEmail":
-            setFieldError("email", error?.description || t("pages.sign_up.errors." + error.code));
+            setFieldError("email", error.description || t("pages.sign_up.errors." + error.code));
             break;
 
           default:
-            if (error?.description) {
-              commonErrors.value.push(error.description);
+            if (error) {
+              commonErrors.value.push(error.description as string);
             }
         }
       }
     }
   }
 });
+
+//todo: remove this workaround when xapi response will be improved to return IdentityErrorType
+function convertToIdentityError(textErrors: string[]): IdentityErrorType[] {
+  const result: IdentityErrorType[] = [];
+  const oneWordRegex = /^\w*$/g;
+  const codeDescriptionRegex = /(\w*): (.*)/gm;
+  textErrors.forEach((errorMessage) => {
+    if (oneWordRegex.test(errorMessage)) {
+      result.push({ code: errorMessage });
+    } else {
+      const matches = codeDescriptionRegex.exec(errorMessage);
+      if (matches && matches.length) {
+        result.push({ code: matches[0], description: matches[2] });
+      } else {
+        result.push({ description: errorMessage });
+      }
+    }
+  });
+
+  return result;
+}
 
 const validateUsernameUniqueness = async (value: string, resolve: (value: boolean) => void) => {
   try {
