+++ client-app/xapi/graphql/account/index.ts
@@ -1,7 +1,4 @@
 export { default as registerAccount } from "./mutations/registerAccount";
-// export { default as createUser } from "./mutations/createUser";
-// export { default as createOrganization } from "./mutations/createOrganization";
-// export { default as createContact } from "./mutations/createContact";
 export { default as updatePersonalData } from "./mutations/updatePersonalData";
 export { default as updateMemberAddresses } from "./mutations/updateMemberAddresses";
 export { default as deleteMemberAddresses } from "./mutations/deleteMemberAddresses";
