+++ client-app/xapi/graphql/account/mutations/registerAccount/requestRegistration.graphql
